pyproject.toml
[tool.poetry]
name = "gaming"
version = "0.1.0"
description = "simple experiment"
authors = ["vladimirgorbachuk@gmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.1"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
